+++ packages/ui/useScroll/Scrollbar.js
@@ -13,11 +13,17 @@ import componentCss from './Scrollbar.module.less';
 const scrollbarTrackHidingDelay = 900; // 900ms + 100ms(fade out duration) = 1000ms.
 
 const addClass = (element, className) => {
-	ReactDOM.findDOMNode(element).classList.add(className); // eslint-disable-line react/no-find-dom-node
+	const node = ReactDOM.findDOMNode(element); // eslint-disable-line react/no-find-dom-node
+	if (node) {
+		node.classList.add(className);
+	}
 };
 
 const removeClass = (element, className) => {
-	ReactDOM.findDOMNode(element).classList.remove(className); // eslint-disable-line react/no-find-dom-node
+	const node = ReactDOM.findDOMNode(element); // eslint-disable-line react/no-find-dom-node
+	if (node) {
+		node.classList.remove(className);
+	}
 };
 
 /*