+++ packages/ui/CHANGELOG.md
@@ -6,7 +6,8 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ### Fixed
 
-- `ui/Scroller` and `ui/VirtualList` to call remaining scrollStopJob function in cleanup phase
+- `ui/Scroller` and `ui/VirtualList` to call `onScrollStop` when scrollbar's visibility changed while scrolling
+
 ## [3.4.6] - 2020-08-24
 
 ### Added